backend/requirements.txt
django==5.0.7
djangorestframework==3.15.2
psycopg[binary]==3.1.18
coverage==7.6.0
freezegun==1.5.1
djangorestframework-simplejwt==5.3.1
black==24.8.0
pylint==3.2.6
celery==5.4.0
django-celery-results==2.5.1